Moved song from gameState to roundState
