+++ src/data_access/InMemoryGameStateGameStateDataAccessObject.java
@@ -5,9 +5,8 @@ import entity.Player;
 import use_case.CheckGuess.CheckGuessGameStateDataAccessInterface;
 import use_case.ReceiveMessage.ReceiveMessageGameStateDataAccessInterface;
 import use_case.SendMessage.SendMessageMainPlayerDataAccessInterface;
-import use_case.SingerChoose.SingerChooseGameStateDataAccessInterface;
 
-public class InMemoryGameStateGameStateDataAccessObject implements SingerChooseGameStateDataAccessInterface, SendMessageMainPlayerDataAccessInterface, CheckGuessGameStateDataAccessInterface, ReceiveMessageGameStateDataAccessInterface {
+public class InMemoryGameStateGameStateDataAccessObject implements SendMessageMainPlayerDataAccessInterface, CheckGuessGameStateDataAccessInterface, ReceiveMessageGameStateDataAccessInterface {
     private final GameState gameState = new GameState();
 
     @Override
