+++ src/data_access/InMemoryRoundStateDataAccessObject.java
@@ -3,11 +3,12 @@ package data_access;
 import entity.RoundState;
 import use_case.CheckGuess.CheckGuessRoundStateDataAccessInterface;
 import use_case.ReceiveMessage.ReceiveMessageRoundStateDataAccessInterface;
+import use_case.SingerChoose.SingerChooseRoundStateDataAccessInterface;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class InMemoryRoundStateDataAccessObject implements CheckGuessRoundStateDataAccessInterface, ReceiveMessageRoundStateDataAccessInterface {
+public class InMemoryRoundStateDataAccessObject implements CheckGuessRoundStateDataAccessInterface, SingerChooseRoundStateDataAccessInterface, ReceiveMessageRoundStateDataAccessInterface {
     private final List<RoundState> rounds = new ArrayList<>();
 
     public RoundState getCurrentRoundState() {
