+++ src/entity/RoundState.java
@@ -5,6 +5,15 @@ import java.util.Map;
 
 public class RoundState {
     private final Map<Player, Boolean> playerGuessStatus = new HashMap<>();
+    private Song song;
+
+    public Song getSong() {
+        return this.song;
+    }
+
+    public void setSong(Song song) {
+        this.song = song;
+    }
 
     public Boolean getGuessStatusByPlayer(Player player) {
         // if the player hasn't been stored in the round yet, default the guess status to false (haven't guessed yet)
