--- src/use_case/SingerChoose/SingerChooseGameStateDataAccessInterface.java
@@ -1,7 +1,0 @@
-package use_case.SingerChoose;
-
-import entity.GameState;
-
-public interface SingerChooseGameStateDataAccessInterface {
-    GameState getGameState();
-}
