+++ src/use_case/SingerChoose/SingerChooseInteractor.java
@@ -1,18 +1,18 @@
 package use_case.SingerChoose;
 
 public class SingerChooseInteractor implements SingerChooseInputBoundary {
-    final SingerChooseGameStateDataAccessInterface singerChooseGameStateDataAccessObject;
+    final SingerChooseRoundStateDataAccessInterface singerChooseRoundStateDataAccessObject;
     final SingerChooseOutputBoundary singerChoosePresenter;
 
-    public SingerChooseInteractor(SingerChooseGameStateDataAccessInterface singerChooseGameStateDataAccessObject, SingerChooseOutputBoundary singerChoosePresenter) {
-        this.singerChooseGameStateDataAccessObject = singerChooseGameStateDataAccessObject;
+    public SingerChooseInteractor(SingerChooseRoundStateDataAccessInterface singerChooseRoundStateDataAccessObject, SingerChooseOutputBoundary singerChoosePresenter) {
+        this.singerChooseRoundStateDataAccessObject = singerChooseRoundStateDataAccessObject;
         this.singerChoosePresenter = singerChoosePresenter;
     }
 
     @Override
     public void execute(SingerChooseInputData singerChooseInputData) {
         // Get the current game state and change the song
-        singerChooseGameStateDataAccessObject.getGameState().setSong(singerChooseInputData.getSong());
+        singerChooseRoundStateDataAccessObject.getCurrentRoundState().setSong(singerChooseInputData.getSong());
         // prepare success view
         SingerChooseOutputData singerChooseOutputData = new SingerChooseOutputData();
         singerChoosePresenter.prepareSuccessView(singerChooseOutputData);
