+++ src/use_case/SingerChoose/SingerChooseRoundStateDataAccessInterface.java
@@ -1,0 +1,7 @@
+package use_case.SingerChoose;
+
+import entity.RoundState;
+
+public interface SingerChooseRoundStateDataAccessInterface {
+    RoundState getCurrentRoundState();
+}
